update status and code index after changing the default branch
