+++ modules/indexer/code/git.go
@@ -30,7 +30,18 @@ func getRepoChanges(ctx context.Context, repo *repo_model.Repository, revision s
 		return nil, err
 	}
 
+	var needGenesis bool
 	if len(status.CommitSha) == 0 {
+		needGenesis = true
+	}
+
+	hasAncestorCmd := git.NewCommand(ctx, "merge-base").AddDynamicArguments(repo.CodeIndexerStatus.CommitSha, revision)
+	stdout, _, _ := hasAncestorCmd.RunStdString(&git.RunOpts{Dir: repo.RepoPath()})
+	if stdout == "" {
+		needGenesis = true
+	}
+
+	if needGenesis {
 		return genesisChanges(ctx, repo, revision)
 	}
 	return nonGenesisChanges(ctx, repo, revision)
